Refactor of components
